+++ changelog.md
@@ -1,4 +1,5 @@
 ## Upcoming
+* Moved Ghostly effect to it's own feature
 * Disable stone tools feature now also disables stone shields from Shields+
 
 ## 5.4.1.2